Includs helper methods for the LanguagePlatform.
The SegmentPair processer offers the possiblity to recover additional information from the ISegmentPair (FileTypeSupport.Framework)
through the LanguagePlatform, that includes the tokenized segments and word counts.

The processor creates a tempoaray translation memory given the settings provided; the TM is then used to recover the LanguagePlatform
segment, which is a sequence of Sdl.LanguagePlatform.Core.SegmentElements.